Move modules to lib
